# Conflicts:
#	Tools/QuickEd/Classes/EditorSettings.cpp
#	Tools/QuickEd/Classes/EditorSettings.h
#	Tools/QuickEd/Classes/UI/mainwindow.cpp
#	Tools/QuickEd/Classes/UI/mainwindow.h


Former-commit-id: 33d2b102f546576b8edbe7fe9aed940a433979a1

+++ Sources/Internal/MemoryManager/MemoryManager.cpp
@@ -35,6 +35,7 @@
 
 #if defined(__DAVAENGINE_WIN32__)
 #include <dbghelp.h>
+#elif defined(__DAVAENGINE_WIN_UAP__)
 #elif defined(__DAVAENGINE_MACOS__) || defined(__DAVAENGINE_IPHONE__)
 #include <execinfo.h>
 #include <dlfcn.h>
@@ -60,30 +61,29 @@
 
 namespace DAVA
 {
-
 struct MemoryManager::MemoryBlock
 {
-    MemoryBlock* prev;      // Pointer to previous block
-    MemoryBlock* next;      // Pointer to next block
-    void* realBlockStart;   // Pointer to real block start
-    void* padding1;         // Padding to make sure that
-    uint32 padding2;        //          struct size is integral multiple of 16 bytes
-    uint32 orderNo;         // Block order number
-    uint32 allocByApp;      // Size requested by application
-    uint32 allocTotal;      // Total allocated size
-    uint32 bktraceHash;     // Unique hash number to identify block backtrace
-    uint32 pool;            // Allocation pool block belongs to
-    uint32 tags;            // Tags block belongs to
-    uint32 mark;            // Mark to distinguish tracked memory blocks
+    MemoryBlock* prev; // Pointer to previous block
+    MemoryBlock* next; // Pointer to next block
+    void* realBlockStart; // Pointer to real block start
+    void* padding1; // Padding to make sure that
+    uint32 padding2; //          struct size is integral multiple of 16 bytes
+    uint32 orderNo; // Block order number
+    uint32 allocByApp; // Size requested by application
+    uint32 allocTotal; // Total allocated size
+    uint32 bktraceHash; // Unique hash number to identify block backtrace
+    uint32 pool; // Allocation pool block belongs to
+    uint32 tags; // Tags block belongs to
+    uint32 mark; // Mark to distinguish tracked memory blocks
 };
 static_assert(sizeof(MemoryManager::MemoryBlock) % 16 == 0, "sizeof(MemoryManager::MemoryBlock) % 16 != 0");
 
 struct MemoryManager::InternalMemoryBlock
 {
-    uint32 allocByApp;      // Size requested by application
-    uint32 allocTotal;      // Total allocated size
+    uint32 allocByApp; // Size requested by application
+    uint32 allocTotal; // Total allocated size
     uint32 padding;
-    uint32 mark;            // Mark to distinguish internal memory blocks
+    uint32 mark; // Mark to distinguish internal memory blocks
 };
 static_assert(sizeof(MemoryManager::InternalMemoryBlock) % 16 == 0, "sizeof(MemoryManager::InternalMemoryBlock) % 16 != 0");
 
@@ -113,7 +113,7 @@ void MemoryManager::RegisterAllocPoolName(uint32 index, const char8* name)
 {
     DVASSERT(name != nullptr && 0 < strlen(name) && strlen(name) < MMItemName::MAX_NAME_LENGTH);
     DVASSERT(index < MAX_ALLOC_POOL_COUNT);
-    DVASSERT(0 == index || allocPoolNames[index - 1].name[0] != '\0');  // Names should be registered sequentially with no gap
+    DVASSERT(0 == index || allocPoolNames[index - 1].name[0] != '\0'); // Names should be registered sequentially with no gap
 
     strncpy(allocPoolNames[index].name, name, MMItemName::MAX_NAME_LENGTH);
     allocPoolNames[index].name[MMItemName::MAX_NAME_LENGTH - 1] = '\0';
@@ -219,7 +219,7 @@ DAVA_NOINLINE void* MemoryManager::Allocate(size_t size, uint32 poolIndex)
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block->realBlockStart));
         block->bktraceHash = 0;
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         if (tlsAllocScopeStack.IsCreated())
         {
@@ -259,7 +259,7 @@ DAVA_NOINLINE void* MemoryManager::AlignedAllocate(size_t size, size_t align, ui
 {
     // TODO: check whether size is integral multiple of align
     assert(size > 0);
-    assert(align > 0 && 0 == (align & (align - 1)));    // Check whether align is power of 2
+    assert(align > 0 && 0 == (align & (align - 1))); // Check whether align is power of 2
     assert(ALLOC_POOL_TOTAL < poolIndex && poolIndex < MAX_ALLOC_POOL_COUNT);
 
     if (align < BLOCK_ALIGN)
@@ -288,7 +288,7 @@ DAVA_NOINLINE void* MemoryManager::AlignedAllocate(size_t size, size_t align, ui
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block->realBlockStart));
         block->bktraceHash = 0;
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         if (tlsAllocScopeStack.IsCreated())
         {
@@ -330,7 +330,7 @@ void* MemoryManager::Reallocate(void* ptr, size_t newSize)
     {
         return malloc(newSize);
     }
-    
+
     MemoryBlock* block = static_cast<MemoryBlock*>(ptr) - 1;
     if (BLOCK_MARK == block->mark)
     {
@@ -350,7 +350,7 @@ void* MemoryManager::Reallocate(void* ptr, size_t newSize)
         return MallocHook::Realloc(ptr, newSize);
     }
 }
-    
+
 void MemoryManager::Deallocate(void* ptr)
 {
     if (ptr != nullptr)
@@ -402,7 +402,7 @@ void* MemoryManager::InternalAllocate(size_t size)
         block->allocByApp = static_cast<uint32>(size);
         block->allocTotal = static_cast<uint32>(MallocHook::MallocSize(block));
         if (0 == block->allocTotal)
-            block->allocTotal = totalSize;
+            block->allocTotal = static_cast<uint32>(totalSize);
 
         // Update stat
         {
@@ -422,7 +422,7 @@ void MemoryManager::InternalDeallocate(void* ptr)
     {
         InternalMemoryBlock* block = static_cast<InternalMemoryBlock*>(ptr) - 1;
         assert(INTERNAL_BLOCK_MARK == block->mark);
-        
+
         // Update stat
         {
             LockType lock(statMutex);
@@ -479,7 +479,7 @@ uint32 MemoryManager::GetTaggedMemoryUsage(uint32 tagIndex) const
 void MemoryManager::EnterTagScope(uint32 tag)
 {
     DVASSERT(tag != 0 && IsPowerOf2(tag));
-    DVASSERT((statGeneral.activeTags & tag) == 0);  // Tag shouldn't be set earlier
+    DVASSERT((statGeneral.activeTags & tag) == 0); // Tag shouldn't be set earlier
 
     {
         LockType lock(allocMutex);
@@ -495,7 +495,7 @@ void MemoryManager::EnterTagScope(uint32 tag)
 void MemoryManager::LeaveTagScope(uint32 tag)
 {
     DVASSERT(tag != 0 && IsPowerOf2(tag));
-    DVASSERT((statGeneral.activeTags & tag) == tag);    // Tag should be set earlier
+    DVASSERT((statGeneral.activeTags & tag) == tag); // Tag should be set earlier
 
     {
         LockType lock(allocMutex);
@@ -542,7 +542,7 @@ void MemoryManager::TrackGpuAlloc(uint32 id, size_t size, uint32 gpuPoolIndex)
     if (iter == gpuBlockMap->end())
     {
         MemoryBlock newBlock{};
-        newBlock.orderNo = id;  // Make use field 'orderNo' as GPU allocation id
+        newBlock.orderNo = id; // Make use field 'orderNo' as GPU allocation id
         newBlock.pool = gpuPoolIndex;
 
         iter = gpuBlockMap->emplace(key, newBlock).first;
@@ -551,7 +551,7 @@ void MemoryManager::TrackGpuAlloc(uint32 id, size_t size, uint32 gpuPoolIndex)
     MemoryBlock& gpuBlock = iter->second;
     gpuBlock.allocByApp += static_cast<uint32>(size);
     gpuBlock.allocTotal = gpuBlock.allocByApp;
-    gpuBlock.mark += 1;     // Make use field 'mark' as number of GPU allocations with given id and pool index
+    gpuBlock.mark += 1; // Make use field 'mark' as number of GPU allocations with given id and pool index
     {
         LockType lock(statMutex);
         UpdateStatAfterGPUAlloc(&gpuBlock, size);
@@ -607,7 +607,7 @@ void MemoryManager::UpdateStatAfterAlloc(MemoryBlock* block, uint32 systemMemory
         statAllocPool[ALLOC_POOL_SYSTEM].allocByApp = systemMemoryUsage;
         statAllocPool[ALLOC_POOL_SYSTEM].allocTotal = systemMemoryUsage;
     }
-    {   // Update total statistics
+    { // Update total statistics
         statAllocPool[ALLOC_POOL_TOTAL].allocByApp += block->allocByApp;
         statAllocPool[ALLOC_POOL_TOTAL].allocTotal += block->allocTotal;
         statAllocPool[ALLOC_POOL_TOTAL].blockCount += 1;
@@ -615,7 +615,7 @@ void MemoryManager::UpdateStatAfterAlloc(MemoryBlock* block, uint32 systemMemory
         if (block->allocByApp > statAllocPool[ALLOC_POOL_TOTAL].maxBlockSize)
             statAllocPool[ALLOC_POOL_TOTAL].maxBlockSize = block->allocByApp;
     }
-    {   // Update pool statistics
+    { // Update pool statistics
         const uint32 poolIndex = block->pool;
         statAllocPool[poolIndex].allocByApp += block->allocByApp;
         statAllocPool[poolIndex].allocTotal += block->allocTotal;
@@ -625,7 +625,7 @@ void MemoryManager::UpdateStatAfterAlloc(MemoryBlock* block, uint32 systemMemory
             statAllocPool[poolIndex].maxBlockSize = block->allocByApp;
     }
 
-    {   // Update tag statistics
+    { // Update tag statistics
         uint32 tags = statGeneral.activeTags;
         if (tags != 0)
         {
@@ -652,18 +652,18 @@ void MemoryManager::UpdateStatAfterDealloc(MemoryBlock* block, uint32 systemMemo
         statAllocPool[ALLOC_POOL_SYSTEM].allocByApp = systemMemoryUsage;
         statAllocPool[ALLOC_POOL_SYSTEM].allocTotal = systemMemoryUsage;
     }
-    {   // Update total statistics
+    { // Update total statistics
         statAllocPool[ALLOC_POOL_TOTAL].allocByApp -= block->allocByApp;
         statAllocPool[ALLOC_POOL_TOTAL].allocTotal -= block->allocTotal;
         statAllocPool[ALLOC_POOL_TOTAL].blockCount -= 1;
     }
-    {   // Update pool statistics
+    { // Update pool statistics
         const uint32 poolIndex = block->pool;
         statAllocPool[poolIndex].allocByApp -= block->allocByApp;
         statAllocPool[poolIndex].allocTotal -= block->allocTotal;
         statAllocPool[poolIndex].blockCount -= 1;
     }
-    {   // Update tag statistics
+    { // Update tag statistics
         uint32 tags = block->tags;
         if (tags != 0)
         {
@@ -686,11 +686,11 @@ void MemoryManager::UpdateStatAfterDealloc(MemoryBlock* block, uint32 systemMemo
 
 void MemoryManager::UpdateStatAfterGPUAlloc(MemoryBlock* block, size_t sizeIncr)
 {
-    {   // Update total statistics
+    { // Update total statistics
         statAllocPool[ALLOC_POOL_TOTAL].allocByApp += static_cast<uint32>(sizeIncr);
         statAllocPool[ALLOC_POOL_TOTAL].allocTotal += static_cast<uint32>(sizeIncr);
     }
-    {   // Update pool statistics
+    { // Update pool statistics
         statAllocPool[block->pool].allocByApp += static_cast<uint32>(sizeIncr);
         statAllocPool[block->pool].allocTotal += static_cast<uint32>(sizeIncr);
         statAllocPool[block->pool].blockCount += 1;
@@ -702,11 +702,11 @@ void MemoryManager::UpdateStatAfterGPUAlloc(MemoryBlock* block, size_t sizeIncr)
 
 void MemoryManager::UpdateStatAfterGPUDealloc(MemoryBlock* block)
 {
-    {   // Update total statistics
+    { // Update total statistics
         statAllocPool[ALLOC_POOL_TOTAL].allocByApp -= block->allocByApp;
         statAllocPool[ALLOC_POOL_TOTAL].allocTotal -= block->allocTotal;
     }
-    {   // Update pool statistics
+    { // Update pool statistics
         statAllocPool[block->pool].allocByApp -= block->allocByApp;
         statAllocPool[block->pool].allocTotal -= block->allocTotal;
         statAllocPool[block->pool].blockCount -= block->mark;
@@ -785,7 +785,7 @@ DAVA_NOINLINE void MemoryManager::CollectBacktrace(Backtrace* backtrace, size_t 
     char8* strPtr = bktraceString;
     size_t bktraceStringLength = 0;
 
-    for (size_t idst = 0, isrc = nskip + 1;idst < BACKTRACE_DEPTH;++idst, ++isrc)
+    for (size_t idst = 0, isrc = nskip + 1; idst < BACKTRACE_DEPTH; ++idst, ++isrc)
     {
         backtrace->frames[idst] = frames[isrc];
 
@@ -834,11 +834,11 @@ void MemoryManager::GetStatConfig(void* buffer, uint32 bufSize) const
     config->bktraceDepth = BACKTRACE_DEPTH;
 
     MMItemName* names = OffsetPointer<MMItemName>(config, sizeof(MMStatConfig));
-    for (uint32 i = 0;i < registeredAllocPoolCount;++i, ++names)
+    for (uint32 i = 0; i < registeredAllocPoolCount; ++i, ++names)
     {
         *names = allocPoolNames[i];
     }
-    for (uint32 i = 0;i < registeredTagCount;++i, ++names)
+    for (uint32 i = 0; i < registeredTagCount; ++i, ++names)
     {
         *names = tagNames[i];
     }
@@ -864,13 +864,13 @@ void MemoryManager::GetCurStat(uint64 timestamp, void* buffer, uint32 bufSize) c
     curStat->statGeneral = statGeneral;
 
     AllocPoolStat* pools = OffsetPointer<AllocPoolStat>(curStat, sizeof(MMCurStat));
-    for (uint32 i = 0;i < registeredAllocPoolCount;++i)
+    for (uint32 i = 0; i < registeredAllocPoolCount; ++i)
     {
         pools[i] = statAllocPool[i];
     }
 
     TagAllocStat* tags = OffsetPointer<TagAllocStat>(pools, sizeof(AllocPoolStat) * registeredAllocPoolCount);
-    for (uint32 i = 0;i < registeredTagCount;++i)
+    for (uint32 i = 0; i < registeredTagCount; ++i)
     {
         tags[i] = statTag[i];
     }
@@ -880,7 +880,7 @@ void MemoryManager::GetCurStat(uint64 timestamp, void* buffer, uint32 bufSize) c
 bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snapshotSize)
 {
     if (lightWeightMode)
-    {   // In lightweight mode snapshot has no sense as it doesn't contains backtraces and symbols
+    { // In lightweight mode snapshot has no sense as it doesn't contains backtraces and symbols
         if (snapshotSize != nullptr)
         {
             *snapshotSize = 0;
@@ -891,8 +891,8 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
     assert(file != nullptr);
 
     const uint32 BUF_SIZE = 64 * 1024;
-    std::vector<uint8, InternalAllocator<uint8>> v(BUF_SIZE);   // For automatic memory management
-    void* buffer = static_cast<void*>(&*v.begin());             // For convenience
+    std::vector<uint8, InternalAllocator<uint8>> v(BUF_SIZE); // For automatic memory management
+    void* buffer = static_cast<void*>(&*v.begin()); // For convenience
 
     const uint32 bktraceSize = sizeof(MMBacktrace) + sizeof(uint64) * BACKTRACE_DEPTH;
 
@@ -906,7 +906,7 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
     if (file->Write(&snapshot) != sizeof(MMSnapshot))
         return false;
 
-    {   // Store memory blocks into file
+    { // Store memory blocks into file
         LockType lock(allocMutex);
 
         const uint32 BLOCKS_IN_BUF = BUF_SIZE / sizeof(MMBlock);
@@ -916,7 +916,7 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
         while (curBlock != nullptr)
         {
             uint32 k = 0;
-            for (;k < BLOCKS_IN_BUF && curBlock != nullptr;++k)
+            for (; k < BLOCKS_IN_BUF && curBlock != nullptr; ++k)
             {
                 MMBlock& dstBlock = destBegin[k];
                 dstBlock.orderNo = curBlock->orderNo;
@@ -933,7 +933,7 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
                 return false;
         }
     }
-    {   // Store function names into file
+    { // Store function names into file
         LockType lock(bktraceMutex);
 
         const size_t SYMBOLS_IN_BUF = BUF_SIZE / sizeof(MMSymbol);
@@ -944,7 +944,7 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
         while (itBegin != itEnd)
         {
             uint32 k = 0;
-            for (;k < SYMBOLS_IN_BUF && itBegin != itEnd;++k)
+            for (; k < SYMBOLS_IN_BUF && itBegin != itEnd; ++k)
             {
                 void* addr = itBegin->first;
                 auto& name = itBegin->second;
@@ -961,7 +961,7 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
                 return false;
         }
     }
-    {   // Store backtraces into file
+    { // Store backtraces into file
         LockType lock(bktraceMutex);
 
         const size_t BKTRACE_IN_BUF = BUF_SIZE / bktraceSize;
@@ -972,13 +972,13 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
         while (itBegin != itEnd)
         {
             uint32 k = 0;
-            for (;k < BKTRACE_IN_BUF && itBegin != itEnd;++k)
+            for (; k < BKTRACE_IN_BUF && itBegin != itEnd; ++k)
             {
                 auto& o = itBegin->second;
 
                 bktrace->hash = o.hash;
                 uint64* frames = OffsetPointer<uint64>(bktrace, sizeof(MMBacktrace));
-                for (size_t i = 0;i < BACKTRACE_DEPTH;++i)
+                for (size_t i = 0; i < BACKTRACE_DEPTH; ++i)
                 {
                     frames[i] = reinterpret_cast<uint64>(o.frames[i]);
                 }
@@ -995,9 +995,9 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
 
     // Write down header
     const uint32 totalSize = sizeof(MMSnapshot)
-                           + sizeof(MMBlock) * snapshot.blockCount
-                           + sizeof(MMSymbol) * snapshot.symbolCount
-                           + bktraceSize * snapshot.bktraceCount;
+    + sizeof(MMBlock) * snapshot.blockCount
+    + sizeof(MMSymbol) * snapshot.symbolCount
+    + bktraceSize * snapshot.bktraceCount;
     if (snapshotSize != nullptr)
     {
         *snapshotSize = totalSize;
@@ -1014,8 +1014,8 @@ bool MemoryManager::GetMemorySnapshot(uint64 timestamp, File* file, uint32* snap
 
 void MemoryManager::SymbolCollectorThread(BaseObject*, void*, void*)
 {
-    const size_t BUF_CAPACITY = 1000;   // Select some reasonable buffer for backtraces
-                                        // to give some job to symbol collector
+    const size_t BUF_CAPACITY = 1000; // Select some reasonable buffer for backtraces
+    // to give some job to symbol collector
     Backtrace* bktraceBuf = new Backtrace[BUF_CAPACITY];
 
     for (;;)
@@ -1029,7 +1029,7 @@ void MemoryManager::SymbolCollectorThread(BaseObject*, void*, void*)
         {
             LockType lock(bktraceMutex);
 
-            for (auto i = bktraceMap->begin(), e = bktraceMap->end();i != e && nplaced < BUF_CAPACITY;++i)
+            for (auto i = bktraceMap->begin(), e = bktraceMap->end(); i != e && nplaced < BUF_CAPACITY; ++i)
             {
                 const Backtrace& bktrace = i->second;
                 if (!bktrace.symbolsCollected)
@@ -1040,7 +1040,7 @@ void MemoryManager::SymbolCollectorThread(BaseObject*, void*, void*)
             }
         }
 
-        for (size_t i = 0;i < nplaced;++i)
+        for (size_t i = 0; i < nplaced; ++i)
         {
             ObtainBacktraceSymbols(&bktraceBuf[i]);
 
@@ -1054,6 +1054,6 @@ void MemoryManager::SymbolCollectorThread(BaseObject*, void*, void*)
     }
 }
 
-}   // namespace DAVA
+} // namespace DAVA
 
-#endif  // defined(DAVA_MEMORY_PROFILING_ENABLE)
+#endif // defined(DAVA_MEMORY_PROFILING_ENABLE)
